+++ BusinessLogicLayer/ModelRequest/Cart/CartRequest.cs
@@ -11,7 +11,7 @@ namespace BusinessLogicLayer.ModelRequest.Cart
     public class CartRequest
     {
         public int TotalItem { get; set; }
-        public double TotalPrice { get; set; }
+        public decimal TotalPrice { get; set; }
         public int AccountId { get; set; }
     }
 
@@ -19,8 +19,8 @@ namespace BusinessLogicLayer.ModelRequest.Cart
     {
         [Range(0, int.MaxValue, ErrorMessage = "Total item cannot be negative.")]
         public int TotalItem { get; set; }
-        [Range(0, double.MaxValue, ErrorMessage = "Total price cannot be negative.")]
-        public double TotalPrice { get; set; }
+        [Range(0, (double)decimal.MaxValue, ErrorMessage = "Total price cannot be negative.")]
+        public decimal TotalPrice { get; set; }
         public int AccountId { get; set; }
         public int? VoucherId { get; set; }
         public ICollection<CartItemRequest> CartItems { get; set; }
