+++ BusinessLogicLayer/ModelRequest/Product/ProductRequest.cs
@@ -17,8 +17,8 @@ namespace BusinessLogicLayer.ModelRequest.Product
         [Required]
         public string? Description { get; set; }
         [Required]
-        [Range(0, double.MaxValue, ErrorMessage = "Product price cannot be negative.")]
-        public double ProductPrice { get; set; }
+        [Range(0, (double)decimal.MaxValue, ErrorMessage = "Product price cannot be negative.")]
+        public decimal ProductPrice { get; set; }
         [Required]
         [Range(0, int.MaxValue, ErrorMessage = "Product quantity cannot be negative.")]
         public int? Quantity { get; set; }
@@ -38,8 +38,8 @@ namespace BusinessLogicLayer.ModelRequest.Product
         [Required]
         public string? Description { get; set; }
         [Required]
-        [Range(0, double.MaxValue, ErrorMessage = "Product price cannot be negative.")]
-        public double ProductPrice { get; set; }
+        [Range(0, (double)decimal.MaxValue, ErrorMessage = "Product price cannot be negative.")]
+        public decimal ProductPrice { get; set; }
         [Required]
         [Range(0, int.MaxValue, ErrorMessage = "Product quantity cannot be negative.")]
         public int? Quantity { get; set; }
@@ -55,7 +55,7 @@ namespace BusinessLogicLayer.ModelRequest.Product
         public int Id { get; set; }
         public string ProductName { get; set; }
         public double Rate { get; set; }
-        public double ProductPrice { get; set; }
+        public decimal ProductPrice { get; set; }
         public int TotalSold { get; set; }
         public List<string>? ImageUrls { get; set; }
     }
@@ -68,7 +68,7 @@ namespace BusinessLogicLayer.ModelRequest.Product
         public int TotalRate { get; set; }
         public int TotalSold { get; set; }
         public string? Description { get; set; }
-        public double ProductPrice { get; set; }
+        public decimal ProductPrice { get; set; }
         public int? Quantity { get; set; }
         public string? MadeIn { get; set; }
         public string? ShipFrom { get; set; }
